Collapse the ai-agents devrig launcher helpers onto devrig-common
The rebase onto PR #444 left two homes for devrig launcher-path logic:
devrig-common's DevrigUserLauncher/HomePaths and the branch-era helpers in
ai-agents (a second DEVRIG_HOME_DIR_NAME spelling '.mcp-steroid',
devrigLauncherFileName, devrigStdioMcpCommand, and the display-path
helpers). ai-agents cannot depend on devrig-common (devrig-common api()s
ai-agents), so the helpers move INTO devrig-common next to
DevrigUserLauncher — which now builds its path via devrigLauncherFileName
instead of spelling the file names a second time.

ai-agents keeps only what is agent-generic: StdioMcpCommand,
stdioMcpServersJson, and the mcp add/remove/list arg builders. The moved
tests go with the code (DevrigUserLauncherTest), plus a new pin for
DevrigUserLauncher.path itself.
