+++ ai-agents/src/main/kotlin/com/jonnyzzz/mcpSteroid/aiAgents/McpServerConfig.kt
@@ -66,76 +66,6 @@ data class StdioMcpCommand(
     val args: List<String> = emptyList(),
 )
 
-/** The directory devrig owns under the user home — the one name every devrig path hangs off. */
-const val DEVRIG_HOME_DIR_NAME = ".mcp-steroid"
-
-/**
- * The stable launcher's file name for this OS: a `.cmd` shim on Windows (so cmd.exe and PowerShell resolve
- * it via PATHEXT), a plain `devrig` script on POSIX.
- */
-fun devrigLauncherFileName(windows: Boolean): String = if (windows) "devrig.cmd" else "devrig"
-
-/**
- * devrig's home rendered for humans — see the display policy on [devrigLauncherDisplayPath].
- */
-fun devrigHomeDisplayPath(userHome: String, windows: Boolean): String =
-    displayPath(userHome, windows, DEVRIG_HOME_DIR_NAME)
-
-/**
- * The stable launcher path rendered for humans.
- *
- * Display policy for every user-visible devrig path: the **real absolute home** (never `~`), joined with
- * the OS-native separator — `C:\Users\me\.mcp-steroid\bin\devrig.cmd` on Windows,
- * `/home/me/.mcp-steroid/bin/devrig` on POSIX. `~` is a lie on Windows (neither cmd.exe nor an
- * `mcp.json`-reading client expands it), and a copy button must put exactly what is displayed on the
- * clipboard — so the display IS the clipboard content, on every OS.
- */
-fun devrigLauncherDisplayPath(userHome: String, windows: Boolean): String =
-    displayPath(userHome, windows, DEVRIG_HOME_DIR_NAME, "bin", devrigLauncherFileName(windows))
-
-/**
- * The one-line command that runs devrig as a stdio MCP server — what a user types into an MCP client that
- * asks for a command line instead of reading an `mcpServers` file. The launcher path follows the
- * [devrigLauncherDisplayPath] policy (real absolute home, OS-native separators) and is quoted when it
- * contains a space, because every shell and client splits an unquoted `C:\Users\First Last\…` in two.
- */
-fun devrigMcpCommandLine(userHome: String, windows: Boolean): String {
-    val launcher = devrigLauncherDisplayPath(userHome, windows)
-    val quoted = if (' ' in launcher) "\"$launcher\"" else launcher
-    return "$quoted mcp"
-}
-
-private fun displayPath(userHome: String, windows: Boolean, vararg segments: String): String {
-    val separator = if (windows) "\\" else "/"
-    // A Windows home can arrive with forward slashes (a config file, a test) — render it Windows-naturally.
-    val home = (if (windows) userHome.replace('/', '\\') else userHome).trimEnd('/', '\\')
-    return (listOf(home) + segments).joinToString(separator)
-}
-
-/**
- * OS-correct stdio invocation of the stable devrig launcher at [launcherPath], with [args].
- *
- * Lives here, in the module both halves already depend on, because two places need the exact same answer:
- * devrig itself when it registers an agent (`DevrigUserLauncher.invocation`), and the IDE plugin when it
- * shows the manual configuration for a client devrig cannot register (Cursor, Windsurf, anything reading an
- * `mcpServers` file). If the two ever built this string separately, the copyable snippet would quietly stop
- * matching what the button writes.
- *
- * Windows runs the `.cmd` through `cmd.exe /d /c` — a `.cmd` is not directly executable as a process image,
- * and `/d` skips any AutoRun script. The launcher path is quoted because `cmd.exe` parses everything after
- * `/c` as ONE command line, so an unquoted `C:\Users\First Last\…` splits and the server never starts.
- * POSIX execs the script directly.
- */
-fun devrigStdioMcpCommand(launcherPath: String, windows: Boolean, args: List<String> = listOf("mcp")): StdioMcpCommand =
-    if (windows) {
-        StdioMcpCommand(
-            command = "cmd.exe",
-            args = listOf("/d", "/c", (listOf("\"$launcherPath\"") + args).joinToString(" ")),
-        )
-    } else {
-        StdioMcpCommand(command = launcherPath, args = args)
-    }
-
 private val stdioMcpServersJsonFormat = Json { prettyPrint = true }
 
 /**
