+++ devrig-common/src/test/kotlin/com/jonnyzzz/mcpSteroid/devrig/DevrigUserLauncherTest.kt
@@ -1,0 +1,100 @@
+/* Copyright 2025-2026 Eugene Petrenko (mcp@jonnyzzz.com); Copyright 2025-2026 JetBrains. Use of this source code is governed by the Apache 2.0 license. */
+package com.jonnyzzz.mcpSteroid.devrig
+
+import org.junit.jupiter.api.Test
+import java.nio.file.Path
+import kotlin.test.assertEquals
+
+class DevrigUserLauncherTest {
+
+    @Test
+    fun `the launcher path is per-OS under the shared bin dir`() {
+        val home = HomePaths(Path.of("/home/u/.mcp-steroid"))
+        assertEquals(Path.of("/home/u/.mcp-steroid/bin/devrig"), DevrigUserLauncher.path(home, windows = false))
+        assertEquals(Path.of("/home/u/.mcp-steroid/bin/devrig.cmd"), DevrigUserLauncher.path(home, windows = true))
+    }
+
+    @Test
+    fun `devrigStdioMcpCommand execs the launcher directly on POSIX`() {
+        val command = devrigStdioMcpCommand("/home/user/.mcp-steroid/bin/devrig", windows = false)
+
+        assertEquals("/home/user/.mcp-steroid/bin/devrig", command.command)
+        assertEquals(listOf("mcp"), command.args)
+    }
+
+    @Test
+    fun `devrigStdioMcpCommand wraps the cmd launcher and quotes a path with spaces`() {
+        // A .cmd is not executable as a process image, so it goes through cmd.exe; /d skips AutoRun.
+        // cmd.exe parses everything after /c as ONE command line, so an unquoted "First Last" would split.
+        val command = devrigStdioMcpCommand("C:\\Users\\First Last\\.mcp-steroid\\bin\\devrig.cmd", windows = true)
+
+        assertEquals("cmd.exe", command.command)
+        assertEquals(
+            listOf("/d", "/c", "\"C:\\Users\\First Last\\.mcp-steroid\\bin\\devrig.cmd\" mcp"),
+            command.args,
+        )
+    }
+
+    // Display policy for user-visible devrig paths: the REAL absolute home (never `~`), OS-native
+    // separators — so what the settings page shows equals what its copy buttons put on the clipboard.
+    // POSIX and Windows are pinned as explicit methods (no parameterized tests in this repo).
+
+    @Test
+    fun `devrigHomeDisplayPath renders the real home on POSIX`() {
+        assertEquals("/home/user/.mcp-steroid", devrigHomeDisplayPath("/home/user", windows = false))
+    }
+
+    @Test
+    fun `devrigHomeDisplayPath renders the real home with backslashes on Windows`() {
+        assertEquals("C:\\Users\\me\\.mcp-steroid", devrigHomeDisplayPath("C:\\Users\\me", windows = true))
+    }
+
+    @Test
+    fun `devrigLauncherDisplayPath names the plain launcher on POSIX`() {
+        assertEquals(
+            "/home/user/.mcp-steroid/bin/devrig",
+            devrigLauncherDisplayPath("/home/user", windows = false),
+        )
+    }
+
+    @Test
+    fun `devrigLauncherDisplayPath names the cmd shim with backslashes on Windows`() {
+        assertEquals(
+            "C:\\Users\\me\\.mcp-steroid\\bin\\devrig.cmd",
+            devrigLauncherDisplayPath("C:\\Users\\me", windows = true),
+        )
+    }
+
+    @Test
+    fun `devrigLauncherDisplayPath normalizes a forward-slash Windows home and a trailing separator`() {
+        assertEquals(
+            "C:\\Users\\me\\.mcp-steroid\\bin\\devrig.cmd",
+            devrigLauncherDisplayPath("C:/Users/me/", windows = true),
+        )
+    }
+
+    @Test
+    fun `devrigMcpCommandLine is the absolute launcher plus the mcp subcommand on POSIX`() {
+        assertEquals(
+            "/home/user/.mcp-steroid/bin/devrig mcp",
+            devrigMcpCommandLine("/home/user", windows = false),
+        )
+    }
+
+    @Test
+    fun `devrigMcpCommandLine leaves a space-free Windows launcher unquoted`() {
+        assertEquals(
+            "C:\\Users\\me\\.mcp-steroid\\bin\\devrig.cmd mcp",
+            devrigMcpCommandLine("C:\\Users\\me", windows = true),
+        )
+    }
+
+    @Test
+    fun `devrigMcpCommandLine quotes a launcher path containing a space`() {
+        // An unquoted "C:\Users\First Last\..." splits into two arguments in every shell and client.
+        assertEquals(
+            "\"C:\\Users\\First Last\\.mcp-steroid\\bin\\devrig.cmd\" mcp",
+            devrigMcpCommandLine("C:\\Users\\First Last", windows = true),
+        )
+    }
+}
