+++ ij-plugin/src/main/kotlin/com/jonnyzzz/mcpSteroid/onboarding/DevrigPromotion.kt
@@ -12,7 +12,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.util.SystemInfo
 import com.intellij.openapi.util.registry.Registry
-import com.jonnyzzz.mcpSteroid.aiAgents.devrigHomeDisplayPath
+import com.jonnyzzz.mcpSteroid.devrig.devrigHomeDisplayPath
 import com.jonnyzzz.mcpSteroid.updates.analyticsBeacon
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
