+++ ij-plugin/src/test/kotlin/com/jonnyzzz/mcpSteroid/settings/McpSteroidConfigurableTest.kt
@@ -9,7 +9,7 @@ import com.intellij.openapi.util.SystemInfo
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.ui.components.fields.ExtendableTextField
 import com.jonnyzzz.mcpSteroid.aiAgents.AiAgentCli
-import com.jonnyzzz.mcpSteroid.aiAgents.devrigMcpCommandLine
+import com.jonnyzzz.mcpSteroid.devrig.devrigMcpCommandLine
 import com.jonnyzzz.mcpSteroid.onboarding.devrigStdioMcpConfigJson
 import com.jonnyzzz.mcpSteroid.onboarding.probeDevrigInstallState
 import com.jonnyzzz.mcpSteroid.server.SteroidsMcpServer
